refactor: decouple auth module functions and organize into service
