+++ src/controllers/authController.go
@@ -1,18 +1,13 @@
 package controllers
 
 import (
-	"fmt"
-	"time"
-
 	"github.com/gofiber/fiber/v2"
-	"github.com/golang-jwt/jwt/v5"
 	"github.com/saahil-mahato/edenmatrix-be/src/models"
-	"github.com/saahil-mahato/edenmatrix-be/src/repositories"
-	"golang.org/x/crypto/bcrypt"
+	"github.com/saahil-mahato/edenmatrix-be/src/services"
 )
 
 type AuthController struct {
-	Repo repositories.AuthRepository
+	Service services.AuthService
 }
 
 func (ac *AuthController) Register(c *fiber.Ctx) error {
@@ -22,17 +17,7 @@ func (ac *AuthController) Register(c *fiber.Ctx) error {
 		return c.Status(fiber.StatusBadRequest).JSON(fiber.Map{"error": "Cannot parse JSON"})
 	}
 
-	hashedPassword, err := bcrypt.GenerateFromPassword([]byte(user.Password), bcrypt.DefaultCost)
-	if err != nil {
-		return c.Status(fiber.StatusInternalServerError).JSON(fiber.Map{"error": "Could not create user"})
-	}
-	user.Password = string(hashedPassword)
-
-	if err := ac.Repo.CreateUser(user); err != nil {
-		return c.Status(fiber.StatusInternalServerError).JSON(fiber.Map{"error": "Could not create user"})
-	}
-
-	return c.Status(fiber.StatusCreated).JSON(fiber.Map{"message": fmt.Sprintf("Successfully created user %s", user.Username)})
+	return ac.Service.CreateUser(c, user)
 }
 
 func (ac *AuthController) Login(c *fiber.Ctx) error {
@@ -42,37 +27,6 @@ func (ac *AuthController) Login(c *fiber.Ctx) error {
 		return c.Status(fiber.StatusBadRequest).JSON(fiber.Map{"error": "Cannot parse JSON"})
 	}
 
-	dbUser, err := ac.Repo.FindUserByEmail(authPayload.Email)
-	if err != nil {
-		return c.Status(fiber.StatusUnauthorized).JSON(fiber.Map{"error": "User is not registered"})
-	}
+	return ac.Service.LoginUser(c, authPayload)
 
-	err = bcrypt.CompareHashAndPassword([]byte(dbUser.Password), []byte(authPayload.Password))
-	if err != nil {
-		return c.Status(fiber.StatusUnauthorized).JSON(fiber.Map{"error": "Invalid credentials"})
-	}
-
-	token := jwt.NewWithClaims(jwt.SigningMethodHS256, jwt.MapClaims{
-		"id":  dbUser.ID,
-		"exp": time.Now().Add(time.Hour * 24).Unix(),
-	})
-
-	var jwtSecret = []byte("your_secret_key")
-	tokenString, err := token.SignedString(jwtSecret)
-	if err != nil {
-		return c.Status(fiber.StatusInternalServerError).JSON(fiber.Map{"error": "Could not generate token"})
-	}
-
-	cookie := fiber.Cookie{
-		Name:     "jwt",
-		Value:    tokenString,
-		Expires:  time.Now().Add(24 * time.Hour),
-		HTTPOnly: true,
-		Secure:   true,
-		SameSite: fiber.CookieSameSiteStrictMode,
-	}
-
-	c.Cookie(&cookie)
-
-	return c.Status(fiber.StatusOK).JSON(fiber.Map{"message": "Login successful"})
 }
